add user entity

+++ src/main/java/com/toy/trelloapi/domain/entity/Card.java
@@ -16,7 +16,7 @@ import java.util.Date;
 public class Card {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer cardId;
+    private Long cardId;
 
     @NotNull
     @ManyToOne
